+++ lib/core/utils/file_util.dart
@@ -1,24 +1,39 @@
 import 'dart:io';
 
 import 'package:dcli/dcli.dart';
-import 'package:mvvm_cli_nerdzlab/core/constatns/pbxproj_constants.dart';
+import 'package:mvvm_cli_nerdzlab/core/constants/pbxproj_constants.dart';
 import 'package:mvvm_cli_nerdzlab/src/dcli/resource/generated/resource_registry.g.dart';
 import 'package:path/path.dart';
 
 abstract class FileUtil {
-  static void copyTemplate(String projectDir) {
-    final outputDirectory = join(Directory.current.path, projectDir);
+  /// Change current directory to directory that under current.
+  ///
+  /// [dir] should exist and be under [Directory.current]
+  static void changeWorkingDirectoryToSubDir(String dir) {
+    final newDirectory = join(Directory.current.path, dir);
+    print(grey('\tChanging to: $newDirectory'));
 
+    Directory.current = newDirectory;
+  }
+
+  /// Copy project template from resource folder.
+  ///
+  /// Any files with .extension will be ignored
+  /// Need to additionally specify in tool/dcli/pack.yaml
+  static void copyProjectTemplate() {
     for (final resource in ResourceRegistry.resources.values) {
-      final localPathTo = join(outputDirectory, resource.originalPath);
+      final localPathTo = join(Directory.current.path, resource.originalPath);
       print(grey('\tUnpacking: ${resource.originalPath}'));
       resource.unpack(localPathTo);
     }
   }
 
-  static Future<void> updatePbxprojData(String projectFolder) async {
-    final pbxprojPath = join(Directory.current.path, projectFolder, 'ios',
-        'Runner.xcodeproj', 'project.pbxproj');
+  /// Update project.pbxproj data.
+  ///
+  /// Will update for dev/prod flavors and schemes.
+  static Future<void> updatePbxprojData() async {
+    final pbxprojPath = join(
+        Directory.current.path, 'ios', 'Runner.xcodeproj', 'project.pbxproj');
     print(grey('\tUpdating: $pbxprojPath'));
     final data = await File(pbxprojPath).readAsString();
 
@@ -38,6 +53,7 @@ abstract class FileUtil {
     await File(pbxprojPath).writeAsString(updatedData);
   }
 
+  /// Remove file via [path]
   static Future<void> removeFile(String path) async {
     final fileToRemove = join(Directory.current.path, path);
     print(grey('\tRemoving: $fileToRemove'));
@@ -45,8 +61,9 @@ abstract class FileUtil {
     await File(fileToRemove).delete();
   }
 
-  static Future<void> updateGitignoreData(String path) async {
-    final fileToUpdate = join(Directory.current.path, path, '.gitignore');
+  /// Updating .gitignore to include .env files
+  static Future<void> updateGitignoreData() async {
+    final fileToUpdate = join(Directory.current.path, '.gitignore');
     print(grey('\tUpdating: $fileToUpdate'));
 
     String data = await File(fileToUpdate).readAsString();
@@ -58,13 +75,19 @@ abstract class FileUtil {
     await File(fileToUpdate).writeAsString(data);
   }
 
+  /// Rename specific identifier in files.
+  ///
+  /// [path] - [Directory.current.path] + path.
+  /// [oldIdentifier] - string to rename.
+  /// [newIdentifier] - new string.
+  /// [allowedFormats] - files with extensions to edit.
   static Future<void> renameIdentifierInDirectory({
-    required String projectFolder,
     required String newIdentifier,
     required String oldIdentifier,
     required List<String> allowedFormats,
+    String? path,
   }) async {
-    final directory = Directory(join(Directory.current.path, projectFolder));
+    final directory = Directory(join(Directory.current.path, path));
 
     // Check if the directory exists
     if (!await directory.exists()) {
@@ -72,7 +95,6 @@ abstract class FileUtil {
     }
 
     // Process all files in the directory and subdirectories
-
     await for (var entity in directory.list(recursive: true)) {
       if (entity is File &&
           allowedFormats.any((format) => entity.path.endsWith(format))) {