+++ lib/core/utils/validator_util.dart
@@ -1,5 +1,9 @@
 import 'package:dcli/dcli.dart';
 
+/// Validator for project name
+///
+/// RegExp: [a-z0-9_]*
+/// For example: 'just_like_this"
 class ProjectNameValidator extends AskValidator {
   const ProjectNameValidator();
 
@@ -9,6 +13,23 @@ class ProjectNameValidator extends AskValidator {
     if (!RegExp(r'^[a-z0-9_]*$').hasMatch(line)) {
       throw AskValidatorException(red(
           'The name should be all lowercase, with underscores to separate words, "just_like_this".Use only basic Latin letters and Arabic digits: [a-z0-9_]'));
+    }
+    return line;
+  }
+}
+
+/// Validator fro yes/no answer
+/// 
+/// RegExp: [YyNn10]
+/// For example: Y/y, N/n, 1/0
+class YesNoValidator extends AskValidator {
+  const YesNoValidator();
+
+  @override
+  String validate(String line, {String? customErrorMessage}) {
+    line = line.trim();
+    if (!RegExp(r'^[YyNn10]$').hasMatch(line)) {
+      throw AskValidatorException(red('The answer should be Y/y, N/n, 1/0'));
     }
     return line;
   }